+++ Specifications/Applications/for_ApplicationArtifactIdentifierStringConverter/for_application_structure_with_4_correct_segments/given/an_application_resource_identifier_converter.cs
@@ -2,7 +2,7 @@ using Dolittle.Artifacts;
 using Machine.Specifications;
 using Moq;
 
-namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter.for_application_with_bounded_context_module_feature_and_recursive_sub_feature.given
+namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter.for_application_structure_with_4_correct_segments.given
 {
     public class an_application_resource_identifier_converter : all_dependencies
     {
