+++ Specifications/Applications/for_ApplicationArtifactIdentifierStringConverter/for_application_structure_with_4_correct_segments/when_converting_string_identifier_missing_application_separator_in_string.cs
@@ -6,15 +6,15 @@ using System;
 using Machine.Specifications;
 using It = Machine.Specifications.It;
 
-namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter.for_application_with_bounded_context_module_feature_and_recursive_sub_feature
+namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter.for_application_structure_with_4_correct_segments
 {
-    public class when_converting_string_identifier_without_application_resource_type_in_string : given.an_application_resource_identifier_converter
+    public class when_converting_string_identifier_missing_application_separator_in_string : given.an_application_resource_identifier_converter
     {
-        static string string_identifier = $"{application_name}#BoundedContext.Module-Resource";
+        static string string_identifier = "Application";
         static Exception exception;
 
         Because of = () => exception = Catch.Exception(() => converter.FromString(string_identifier));
 
-        It should_throw_missing_application_resource = () => exception.ShouldBeOfExactType<MissingApplicationArtifactType>();
+        It should_throw_unable_to_identify_application = () => exception.ShouldBeOfExactType<UnableToIdentifyApplication>();
     }
 }
