+++ Specifications/Applications/for_ApplicationArtifactIdentifierStringConverter/for_application_structure_with_4_correct_segments/when_converting_string_identifier_with_bounded_context_and_module.cs
@@ -3,12 +3,14 @@
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 using System.Linq;
+using Dolittle.Strings;
 using Machine.Specifications;
+using Moq;
 using It = Machine.Specifications.It;
 
-namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter.for_application_with_bounded_context_module_feature_and_recursive_sub_feature
+namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter.for_application_structure_with_4_correct_segments
 {
-    public class when_converting_string_identifier_with_only_bounded_context : given.an_application_resource_identifier_converter
+    public class when_converting_string_identifier_with_bounded_context_and_module : given.an_application_resource_identifier_converter
     {
         const string bounded_context_name = "TheBoundedContext";
         const string module_name = "TheModule";
@@ -17,10 +19,11 @@ namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringCon
         const string second_level_sub_feature_name = "TheSecondLevelSubFeature";
         const string resource_name = "MyResource";
         const int artifact_generation = 1;
-
+        
         static string string_identifier =
             $"{application_name}{ApplicationArtifactIdentifierStringConverter.ApplicationSeparator}" +
             $"{bounded_context_name}{ApplicationArtifactIdentifierStringConverter.ApplicationLocationSeparator}" +
+            $"{module_name}" +
             $"{ApplicationArtifactIdentifierStringConverter.ApplicationArtifactSeparator}{resource_name}"+
             $"{ApplicationArtifactIdentifierStringConverter.ApplicationArtifactTypeSeparator}{artifact_type_name}"+
             $"{ApplicationArtifactIdentifierStringConverter.ApplicationArtifactGenerationSeperator}{artifact_generation}";
@@ -32,7 +35,9 @@ namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringCon
         It should_return_a_matching_identifier = () => identifier.ShouldNotBeNull();
         It should_hold_the_application = () => identifier.Application.ShouldEqual(application);
         It should_hold_the_artifact = () => identifier.Artifact.Name.Value.ShouldEqual(resource_name);
+        It should_hold_the_two_segments = () => identifier.Location.Segments.Count().ShouldEqual(2);
         It should_hold_the_bounded_context_segment = () => identifier.Location.Segments.First().Name.AsString().ShouldEqual(bounded_context_name);
+        It should_hold_the_module_segment = () => identifier.Location.Segments.ToArray()[1].Name.AsString().ShouldEqual(module_name);
         It should_hold_the_artifact_type = () => identifier.Artifact.Type.ShouldEqual(artifact_type.Object);
         It should_hold_the_artifact_generation = () => identifier.Artifact.Generation.Value.ShouldEqual(artifact_generation);
     }
